Add language-specific font configuration
meta.xml can now have a <font> tag, which gives the name of the font
that the language needs. This will directly control the interface font
when the language is active, and will soon also control the font used
for each option on the language screen.

+++ desktop_version/src/Textbox.cpp
@@ -105,7 +105,7 @@ void textboxclass::resize(void)
     for (size_t iter = 0; iter < lines.size(); iter++)
     {
         int len = font::len(print_flags, lines[iter]);
-        if (len > (unsigned int)max) max = len;
+        if (len > max) max = len;
     }
 
     // 16 for the borders
